requirements.txt
# Este archivo contiene las dependencias de PRODUCCIÓN.
# Son las mínimas necesarias para que la aplicación funcione.

aiofiles==24.1.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
google-cloud-translate==2.0.2
Faker==26.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
mdurl==0.1.2
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
psycopg[binary]==3.1.19
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydicom==2.4.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
slowapi==0.1.9
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1